Remove request-promise from submission apis
Replaces request-promise with built in XMLHTTPRequest functionality.
This fixes a bug where the app could not submit batches or payloads over
TLS. This is because protocol had to be included in the URL for
request-promise, which was hard coded as http. Removing the dependency
allows us to just use a relative URL which will get handled by the proxy

Signed-off-by: Darian Plumb <dplumb@bitwise.io>