Update gameroom compose files to use certs and keys generated with cli
This removes the use of the --generate-cli and replaces it with
generate the certificates and keys with the splinter cli. Also adds
the --insecure flag that is required when using generated certs.

The key registry volume is replaced with a gameroom_shared volume that
will store the signing keys, generate certificates and certificate keys.
The generate-key-registry image is renamed to generate-shared.

Also updates the README to point to the new image name and path.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>