Add connection manager daemon
Add connection manager repl server. This is the server side of a toy
application meant to showcase the functionality of the connection
manager.

The rest api interface is 1:1 with the connection manager's
functionality.

`POST /connections/create`

```
{
  "url": <url>
}
```

Creates a new connection with <url>

`DELETE /connections/delete`

```
{
  "url": <url>
}
```

Removes the a connection to <url>

`GET /connections/fetch`

Lists all active connections

Signed-off-by: Ryan Banks <rbanks@bitwise.io>