Catch and handle error in getNode helper API call
When listing proposals, gamerooms, or notifications from the UI, we have
to make a separate call to the /nodes endpoint to get information from
the node registry. If one of the nodes referenced in a list of resources
was removed or renamed, this would cause an error and none of the items
would be displayed.

This commit catches the error and logs a warning so that it does not
break other resources. It also replaces the missing node info with some
default values.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>