Only validate signers against local keys
If a proposal or vote is submitted on behalf of the node, validate that
the signing public key is allowed to make the request.

This change alters the behavior, such that the key registry is not
required to be the same across all nodes, but is only a local concern of
the node. This is better from both a security perspective, as well as a
deployment perspective.  It is more secure in that the keys are managed
by the node's administrator only.  It is easier to deploy, in that the
keys only need to be deployed with a single node, not shared in some
out-of-band way with each member of a network.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>