Move CredentialsStore to credentials::store
Move the CredentialsStore trait to the credentails::store module, and
the diesel implementation from credential_store to
credentails::store::diesel. This move is an initial change towards

a) Supporting the new model for alternative diesel backends that was
introduced in the UserStore code
b) Allows for better organization of alternative store implementations,
such as in-memory, etc.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>