Check circuit ID before listing games
This fixes a bug where games were disappearing on refresh. In Vuex,
getters are observer objects, so just checking if the object is not null
is not sufficient to know if there is actually a gameroom selected or
not. This was causing games to be fetched for a gameroom called
'undefined'. Checking if the circuit_id property is defined fixes this
issue.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>