Replace spinners with loader component
Instead of having seperate spinners that are configured in each
component, this commit adds a Loader component that gets rendered based
on the isLoading state. The router will now commit a relevant message
and state when a new page is being loaded.

This refactor also fixes a bug where the previous spinner was pushing
out the margins of the page and causing scrollbars to flash.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>