Refactor authorize_user function and update affected routes
Refactors the authorize_user function to remove the user_id argument,
add a validation argument, and return the claims. This will allow for
flexibility in configure validation criteria and allow for
authorize_user to be used by routes that only have access to the user_id
via the claims in the authorization header.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>