Have login route return a refresh token
The login route now returns a refresh token in addition to the access
token. The refresh token is to be kept by the client in local storage.
When the client's access token expires it can send a request containing
the refresh token to `/biome/tokens` to obtain a new access token.

Sample request

POST /biome/tokens
   {
        "refresh_token": <refresh token for requesting an auth token>
   }

Signed-off-by: Ryan Banks <rbanks@bitwise.io>