Add /biome/token endpoint
The `/biome/token` takes a payload containing a refresh token. It then
checks that the user has a valid but potentially expired authentication
token, then checks to see if it has a valid refresh token for the
provided user using the token store, if it does it validates that the
refresh token supplied is valid and issues the user a new authentication
token. If the refresh token is no longer valid it removes the user's
refresh token from the database.

Sample Payload

The payload should be in the JSON format:
{
  "refresh_token": <refresh token for requesting an auth token>
}

Endpoint returns a payload containing a new auth token
{
  "token": <new auth token>
}

Signed-off-by: Ryan Banks <rbanks@bitwise.io>