Make add, remove, and update functions take a mutable reference to self.
While writing a mock user object for testing I noticed that the current
interface for user store limits what data types can be used to back a
store implementation. Any implementation that mutates the struct
implementing UserStore is not allowed because the trait only take a
mutable reference to self. In order to make the UserStore trait more
flexible, this commit  updates `remove_user`, `update_user` and
`add_user` to take a mutable reference to self.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>