Add generic Request struct to replace HttpRequest
Introduces a generic `Request` struct for the splinter REST API that
encapsulates HTTP request data such as the path and query parameters and
HTTP headers. This struct replaces the `actix_web::HttpRequest` struct
in the REST API code used throughout libsplinter.

The actix request struct is now converted to the generic struct which is
then passed to the `Resource` implementations. This conversion is
defined in a new `splinter::rest_api::actix_impl` module, which will be
used for other actix-specific REST API implementations. This module
allows for a clean division between the generic REST API code and the
actix-specific REST API code.

The `Request` struct is built with a `RequestBuilder` which will enable
easy backwards compatibility.

Signed-off-by: Logan Seeley <seeley@bitwise.io>