Replace `ResponseError` with `WebSocketError`
This error is currently only used when starting a WebSocket, so it is
renamed to better fit its use. It's also converted to a struct with
context and optional source rather than an enum to match errors
elsewhere in splinter.

Note that the actix `Error` that is converted to a `WebSocketError` in
the `new_websocket_event_sender` function is added to the context string
rather than included as the `source` error; this is because actix's
`Error` does not implement `std::error::Error`, so it cannot be used as
a source.

Signed-off-by: Logan Seeley <seeley@bitwise.io>