Add initial ServiceConnectionManager
Add the initial ServiceConnectionManager implementation which can accept
incoming connections from a connection manager and store the information
about the connection in its internal state.

A ServiceConnectionManager is created via a builder, which is used to
configure and start the background thread.  Using this method, a
ServiceConnectionManager cannot be improperly configured.

The ServiceConnectionAgent runs on the background thread and receives
messages from the ServiceConnection manager.  This agent owns the state,
stored in a ServiceConnectionMap.  Changes occur by processing the
messages in the order of receipt.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>