Add node-file arg to splinter circuit create cmd
Add the `--node-file` argument to the `splinter circuit create` command.
This argument, which can be used instead of or in conjunction with the
`--node` argument, allows the user to load a list of nodes from a .yaml
file either on the local file system or from a remote server with HTTP.
This argument supports the existing node file types used in splinter,
namely those used by the node registry and the `splinter node alias`
commands.

Signed-off-by: Logan Seeley <seeley@bitwise.io>