Display new circuit for splinter circuit create
Updates the `splinter circuit create` command to display the newly
proposed circuit after it is submitted. This allows the user to
immediately see the circuit that was generated without having to run
`splinter circuit proposals`.

Also reorganizes the code slightly and updates the dry-run output to be
of the same format.

Signed-off-by: Logan Seeley <seeley@bitwise.io>