Add update_keys_and_password method
Adds a method to update the keys for an existing user as well
as the password for the associated user. This is assuming that
if a user's password has changed, the associated key pairs, which
hold an encrypted private key, that will also need to be updated.

This also makes the biome::credentials::store::diesel::schema module
public within the biome module to be used in this method.

Signed-off-by: Shannyn Telander <telander@bitwise.io>