Aggressively return WouldBlock in WsTransport
The initial approach taken to WsTransport was to make it as blocking as
possible. That may be an incorrect approach. This changes makes it as
non-blocking as possible, which should be friendlier to use with poll().

This is a separate commit so that it is easy to revert for testing
purposes.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>