Copy splinterd files in gameroomd dockerfile
This is a workaround for compiling in the splinter-dev image. During
builds, the workspace Cargo.toml makes all splinter components
recompile. As various features are removed or renamed, splinterd's
Cargo.toml becomes out of date, depending on features that are no longer
available. Copying over the updated files from the PR fixes the broken
dependencies.

Signed-off-by: Logan Seeley <seeley@bitwise.io>