Replace "database" feature with mix of diesel, postgres
Users of libsplinter should use the "postgres" feature (or in the
future, the "sqlite" feature) instead of an intermediate layer.

Some code that was protected by "database" is now protected by "diesel"
instead. This code should be enabled indirectly by enabling "postgres".

In serveral places in the biome module, diesel-specific code was not
guarded by a feature; this fixes that as well.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>