Mark required arg for splinter circuit show cmd
Updates the `circuit` argument for the `splinter circuit show` CLI
command to be required by Clap. This argument was already required by
the corresponding `Action` code, it just wasn't marked as required in
the Clap code. This change allows the absence of the argument to be
detected earlier and a helpful error message displayed by clap instead
of a simple error.

Signed-off-by: Logan Seeley <seeley@bitwise.io>