Shutdown REST API on test completion
When testing REST API endpoints, it is necessary to shutdown the REST
API server, so that all the in-use sockets are closed.  Without this
change, tests will fail due to "Too many open files", particularly on
MacOS, but may begin to occur on linux, etc, with the addition of more
REST API tests.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>