Move database migrations into biome/migrations
Each major component (Biome, etc) that has database migrations should
be able to run those migrations independently of any other component.
Prior to this commit, both splinter::database::run_migrations and the
biome migrations had to be run in order.  This commit moves the initial
migration to biome, and removes the single migration in the database
module.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>