Improve builder error messages in circuit CLI
Updates the error messages displayed by the splinter circuit CLI when an
error is returned by one of the admin message builders. Specifically,
removes mentions of internal splinter structs and the term "field",
which doesn't have meaning to the user.

Also includes the service ID in the error messages for service builds
that fail.

Signed-off-by: Logan Seeley <seeley@bitwise.io>