Add NetworkMessageSendQueue and NetworkMessageSender
Instead of having a Sender<SendRequest>, dispatch handlers and
other components that need to send a message over the network
will get a NetworkMessageSender. The sender will hide the
implementation details about sending a message to the network
object. As a result SendRequest is now private.

This change is also in support of removing atomic bools used in
shutdown. The NeworkMessageSendQueue now provides a ShutdownSignaler
that will notify the sending thread to shutdown.

This change required that all handler test were rewritten to use
networks and inproc connections. The authorization handlers required
tcp connections because inproc connections are automatically authorized.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>