Check presence of service/template args with clap
Adds the `required_unless("service")` modifier to the `--template`
argument for the `splinter circuit propose` command. `--service` is
already `required_unless("template")`. One of these two arguments is
required for specifying the services on the circuit.

This change allows clap to detect that one of these arguments is
provided, which allows us to remove the check from the `Action` code.

Signed-off-by: Logan Seeley <seeley@bitwise.io>