Check for dup service args in circuit propose CLI
Updates the `splinter circuit propose` command's circuit builder to
check for duplicate service arguments when a new argument is added and
return an error if one is found.

Before this change, the
`CreateCircuitMessageBuilder::apply_service_arguments` method only
checked for duplication of the "peer_services" argument. All arguments
should be checked for duplication to prevent the user from
unintentionally overwriting arguments.

Signed-off-by: Logan Seeley <seeley@bitwise.io>