Improve err checking/msgs for circuit CLI actions
Improves the error messages in the circuit CLI actions and circuit
builder to be more helpful, concise, and not expose internal
implementation details or useless information.

Also adds some additional error conditions to the argument parsing
functions of the `splinter circuit propose` command, like checking for
empty or missing strings.

Signed-off-by: Logan Seeley <seeley@bitwise.io>