Add MeshShutdownSignaler and update Mesh to shutdown
This commit updates InternalEnvelope to be an Enum with
Message and Shutdown options. Mesh will now has a function
that will return a MeshShutdownSignaler that can be used
to shutdown Mesh.

A function has also been added to Network to return a
MeshShutdownSignaler as well as Network is just a wrapper
around Mesh.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>