Add DispatchLoopBuilder and implement shutdown
Before a DispatchLoop would need to be run in a separate thread
that splinterd would need to create. The loop would take an
Arc<AtomicBool> to signal shutdown and a recv to get messages from.

Now a DispatchLoopBuilder will take the dispatcher and start up its
own channel and thread on build. The DispatchLoop provides methods
to get a sender and a shutdown_signaler. This change also
required that the DispatchMessage got changed into a enum with
Message and Shutdown.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>