Add initial network test for splinter
This test simulates a brief network interruption between nodes with a created
circuit which causes the splinter nodes to disconnect. When the interruption is
resolved, a second circuit is proposed and created to verify reconnection.

The setting "net.ipv4.tcp_retries2=3" in the Compose file is used to limit the
number of retries before the TCP stack will inform the layers above that the
connection has been broken. tcp_retries2 uses an exponential backoff starting at
approximately 200ms, so a tcp_retries2 value of "3" should cause a timeout
after about 1.4 seconds.

More detailed information about the network settings can be found at
https://www.kernel.org/doc/Documentation/networking/ip-sysctl.txt

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>