Remove readiness argument from test_poll()
This reverts the changes made in 9d57fea22.

The readiness state of the polling is defined by the test itself and
should not change based on the transport.

Note that this test is fragile as it expects the readiness state to
immediately be readable, which is not necessarily how the events will
be delivered.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>