Convert Handler from generics to associated types
Switch the dispatch Handler trait from using generics to using
associated types.  This alters some constraints on the Handler, in that
it can only be implemented once per type.  It will improve type use
later, for increased parameters to a Handler's type constraints.

This change removes the trait implementations for closures or functions,
as this was a little-used feature.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>