Add extra clone methods to RwNodeRegistry
Adds the `clone_box_as_reader` and `clone_box_as_writer` methods to the
`RwNodeRegistry` trait (and all implementations) for cloning a registry
as a `NodeRegistryReader` and `NodeRegistryWriter`, respectively. These
methods will help coerce the RW registry (which is essentially just a
marker trait for reader + writer) into the two "sub" traits.

Signed-off-by: Logan Seeley <seeley@bitwise.io>