Add `registries` config to splinterd
Adds the `registries` config option to splinterd for specifying
read-only registries to add to the unified node registry. Currently,
only file-base read-only registries are supported; they must be a valid
path to a local file, prefixed with the `file://` URI scheme. In the
future, `http://` and `https://` URIs will be supported with a remote
node registry.

Signed-off-by: Logan Seeley <seeley@bitwise.io>