Add MessageSender trait
Add a message sender trait that can be linked to a particular
source/destination type.  This trait can allow messages to be sent on
the peer level or on the connection level, depending on how source is
configured for the dispatcher instance.

Additionally, adds a private module, `dispatch_peer`, to hold the
implementation required for the `NetworkMessageSender`.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>