Add OPTIONS check if cors enabled
Adds a step to check if a request has method OPTIONS, as this
indicates a preflight check automatically sent by the client
before the actual request is sent to the server to verify that
cors is enabled. This request should then be responded to with
the correct headers and a 200 response, rather than sending the
request straight to the resource as this makes any errors returned
from the resource seem as though the preflight check is the source
of the error.

Signed-off-by: Shannyn Telander <telander@bitwise.io>