Make source an option for StorageErrors
The source field for `<store_error>::StorageError` has been changed to
`Option<Box<dyn Error>>` this is to make it easier to handle cases where
the underlying error isn't interesting or important to surface, or if
the error doesn't implement the `Error` trait. For example if the underlying
error is a String or an Enum. Another benefit to this change is it makes
it easier to work with errors that have generics type that can be
awkward to work with like `std::sync::PoisonError<T>`.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>