Adds justfile for simple cross-repo build, lint, test
Just is a tool for running commands. See:

    https://github.com/casey/just

In this case, the justfile allows you to use the command to run:

  just build
  just lint
  just test

These will operate across the entire repository.

The intent is to use Just outside of docker during development, while
providing reasonable coverage of the commands normally run inside docker
during the PR checks. The behavior of the justfile and docker CI are not
guaranteed to be equivalent.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>