Refactor splinterd local registry location config
Renames the `local_node_registry_location` argument for splinterd to
`node_registry_directory` and changes it to be the directory the local
node registry will be located in (the splinter state directory). This
directory will also be used for the local caches of the remote node
registries.

Also renames the local node registry file from `nodes.yaml` to
`local_registry.yaml` to distinguish it from the remote node registry
files.

Signed-off-by: Logan Seeley <seeley@bitwise.io>