Remove --transport option from splinterd
This commit removes the --transport option from splinterd.

Splinterd will now create a MultiTransport with all configured
transports added. Connections and listeners created are now based
on the protocol prefix in the endpoint.

For example tcps://ip:port will set up a TLS listener and
tcp://ip:port will set up a TCP listener.

Splinterd will try to include a TlsTransport by default. If the
certificates and keys are not available splinterd will error. To
turn TLS off, a new flag has been added --no-tls.

If a protocol prefix is not provided with the endpoint it will default
to TCP.

This commit also renames ADMIN_SERVICE_ADDRESS to INTERNAL_SERVICE_ADDRESS
to more accurately describe the connection being created.

Also fixes a bug where the health service would never be created because
it was placed after the main_loop.join.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>