Implement `splinter registry build` command
Adds the `splinter registry build` command to the splinter CLI. This
command queries a splinter node's `GET /status` endpoint for details
about the node and combines them with public keys read from the
filesystem and provided metadata values to generate a node registry
entry and insert it into a node registry YAML file.

This command is primarily intended for example applications like
gameroom.

Signed-off-by: Logan Seeley <seeley@bitwise.io>