Update default for advertised_endpoints config
Updates the `advertised_endpoints` config for splinterd to default to
the value of `network_endpoints` if no value is provided. This is
preferable to defaulting to an empty list, because the intent of
`advertised_endpoints` is to mirror the network endpoints except in the
case that the endpoints visible to other daemons is different from what
splinterd binds to.

Signed-off-by: Logan Seeley <seeley@bitwise.io>