Add simple thread pool
Add a simple thread pool implementation for use with authorization.
Authorization will not be able to use its own mesh (due to limitations
with epoll/kqueue, etc) , so it requires reading from connections on
their own threads.  This pools will allow for multiple threads to be
used, but up to a limit.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>