Add connection authorization pool
Connections are added to the authorization pool, which handles the
authorization state changes for each connection.  Once a connection has
completed authorization, the connection is removed from the pool and its
status is returned to a provided callback.

Each connection's authorization processes is added as a job to the
underlying thread pool of the authorization pool.  This allows for
multiple connections to be authorized concurrently.

There are several major changes in this authorization code over the
existing authorization code used by the previous network stack:

- there is no special operations for inproc connections.  This should be
  left to the creator of the connection to decide if they require
  authorization or not.
- the connection is owned by the pool until it has completed the
  authorization handshake.
- authorization is not based on peer identity, but on a generic idea of
  identity.  This identity can be passed to the pool based on higher-
  level uses of the connection manager.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>