Update remote registry thread sleep time
Updates the remote registry's automatic refresh thread to check for
shutdown every second while waiting for the refresh period to end.
Previously, the thread would only check for the shutdown signal after
the refresh period; this would is problematic if the refresh period is
long because it would delay the shutdown. With this change, the thread
will be able to react to the shutdown signal within a second.

Signed-off-by: Logan Seeley <seeley@bitwise.io>