Update example applications for new registry
Updates the example applications to reflect the changes to the node/key
registry.

For most examples, the only changes necessary are to add a `keys` field
to the nodes in the node registry and remove the key registry.

For the gameroom example, the node registry file is generated with the
`splinter registry build` command, the file is served up by the new
`registry-server` docker container, and the splinterd nodes have the
file added as a remote registry.

Additionally, the gameroom example has been updated to use private keys,
truncated to 6 characters, to represent XO game players. With the
removal of the key registry, gameroom does not currently have a way to
associate keys with friendly player names. This will be addressed in a
future PR.

Signed-off-by: Logan Seeley <seeley@bitwise.io>