Include identity in connection results
When requesting or adding an inbound connection, the identity that has
been returned based on the authorization handshake is included in either
the result of the request or on the notification for the new inbound
connection.  This allows layers, such as the peer manager or service
connection manager to make use of these identities.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>