Update registry list iterator to be exact size
Updates the return type of the `NodeRegistryReader::list_nodes` method
to use an `ExactSizeIterator` instead of just `Iterator`. The current
implementations of the trait already know the size of the returned
iterator, so this has no impact besides a bit of code reorganization.
The size of the iterator will be used by the `GET /admin/nodes` REST API
endpoint to determine paging (just like the `GET /admin/circuits`
endpoint).

Also cleans up the `list_nodes` definition and implementations by using
the `NodeIter` type for all of them.

Signed-off-by: Logan Seeley <seeley@bitwise.io>