Accept temporary peers
Accept incoming connections as temporary peers.  This state for the peer
means that an remote node has peered with the local node, but that
remote peer is not yet referenced locally for any reason.

Once the local peer has referenced the peer id, the peer is promoted
into the reference map and can participate in the fallback endpoint
cascade.

For example, it is not yet in any circuit relationship yet - either via
negotiating a circuit or an existing circuit. In either of those
circumstances, the admin service will add a peer reference, which would
promote the peer connection.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>