Update `get_file_path` function
Renames the `get_file_path` function to `get_tls_file_path` to
more accurately reflect the files edited within this function.

This also updates the function to check if the file given
is an absolute path, or begins with `./` or `../`, in which
case the tls_cert_dir value will not be appended to the file.
If none of these are true, the tls_cert_dir is used as the
prefix to the file.

This also updates the ConfigBuilder to check every file,
regardless of the ConfigSource with the `get_tls_file_path`
function.

Signed-off-by: Shannyn Telander <telander@bitwise.io>