Allow dead-code for registry config options
This allows dead code for the registry-remote options. This prevents
warnings when the registry-remote feature is not enabled. An alternate
approach would be to properly guard code with the registry-remote
feature; however, this is more intrusive, so taking the easy path here
(which is used for many other config options as well).

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>