Update 'just lint' to run on all crates and exit on warnings
Previously, the lint rule was not properly looping for all crates. This
fixes that. It also adds -D warnings which causes compilation warnings
to be fatal.

Also adds a 'clean' rule and links 'lint' to 'clean'; without this,
clippy does not always error. In the future, we may want this to be
controlled with a just option.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>