Fix scabbard get_state_with_prefix not found error
Updates the `ScabbardState::get_state_with_prefix` method to check the
type of any returned errors and return an empty iterator in the event of
a `NotFound` error. Returning an empty iterator when no entries are
found under the given prefix is the intended behavior of this function,
not returning an error; it's valid to query for state entries under a
prefix with no set addresses.

Signed-off-by: Logan Seeley <seeley@bitwise.io>