Add check for SPLINTER_HOME env var
Adds a check when building the default directory in the
CLI certs actions. This will use the value of the SPLINTER_HOME
environment variable, if set. If another environment variable
is used for the directory paths, the SPLINTER_HOME value is
disregarded.

This also adds a check within splinterd, to see if this value
is set. If it is, the appropriate directory name is appended
to the end of the SPLINTER_HOME variable. Other environment
variables, such as SPLINTER_CERT_DIR and SPLINTER_STATE_DIR,
have precedence over SPLINTER_HOME.

Also adds a check when the state directory is used in the
libsplinter admin service.

Signed-off-by: Shannyn Telander <telander@bitwise.io>