Move the scabbard CLI to its own crate
Moves the scabbard CLI to its own crate. We anticipate that it will be
easier to maintain if it is separate from the scabbard crate.

Also adds `stable` to the `experimental` feature for both the new CLI
crate and the existing scabbard crate, since this is how features are
configured in all other splinter crates.

Signed-off-by: Logan Seeley <seeley@bitwise.io>