Add initial ServiceInterconnect
The service interconnect mirrors the peer interconnect, but operates on
services, service identities and connection ids.  Messages, in this case
are Component Messages.

Unlike the ServiceConnectionManager and PeerManager, it differs far less
in its internals, and there is a likelyhood that the two could be merged
into a single interconnect in the future.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>