Refresh local YAML registry when file is modified
Updates the `LocalYamlNodeRegistry` to refresh is in-memory cache on
read if the backing file has been modified since the last read. Before
this change, updates to a read-only local YAML registry would not be
reflected like updates to a read-only remote YAML registry. This will
allow local registries to be updated as necessary without restarting.

Signed-off-by: Logan Seeley <seeley@bitwise.io>