Log service start failures
Prior to this commit, when the admin service would fail to restart a
service, it would propagate the error up. This would result in the admin
service failing to startup. By logging the service startup error, the
admin service can continue to function normally, even if a service
fails to start.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>