Separate registry REST API from admin
Separates the Splinter registry's REST API endpoints from the admin
service's by moving the `/admin/nodes` endpoints to `/registry/nodes`
and adding a `REGISTRY_PROTOCOL_VERSION` to handle versioning of the
registry's endpoints.

The registry is not a sub-component of the admin service, so this
organization more accurately represents the registry's place in the
Splinter ecosystem.

Signed-off-by: Logan Seeley <seeley@bitwise.io>