Move database and database-migrate-biome to default
Moves the database feature and the database-migrate-biome feature to the
default features list. We are moving database-migrate-biome into default
because we don't expect users to recompile the cli binary with the
stable feature flag. Additionally, we are moving database into stable
because its corresponding libsplinter feature, postgres, was recently
stabilized.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>