Start including Node id in SplinterNode Serialize
This will now include the splinter node id on serialize
and deserialize. Before the node id would be removed
but now it will be included and if not present it will
be filled in with an empty string for backwards
compatibility.

This is a similar fix to include circuit id that went
in previously.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>