Add missing serde decorator for BatchStatusResponse
The following decorator was missing from the local structs
used in the Scabbard RestAPI and ScabbardClient.

This decorator exists on the structs used in the scabbard
services. Without it, gameroom was showing an error toast
on submitting a XO transaction, even though the transaction
would be accepted and show up on the game board.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>