Add function to calculate node_id
Adds a function to check if a node_id is saved within a file,
and if it is, verifies that it matches the node_id loaded into
the config object, if the node_id option is present. If the
node_ids do not match, an error is returned.

Otherwise, uses the config node_id or a randomly generated node_id
and creates the node_id file.

Signed-off-by: Shannyn Telander <telander@bitwise.io>