Move crate::{matrix -> transport::matrix}, prefix with ConnectionMatrix
This changes Matrix to ConnectionMatrix. The matrix is essentially a
switching/routing mechanism from id to Connection, so tying it to
Connection closely seems appropriate.

Moved to transport::matrix because it is so tightly coupled with
Connection and doesn't introduce any new dependencies or relationships
outside of the transport module. This makes it feel well-aligned with
the transport module.

This changes 'Envelope' to 'ConnectionManagerEnvelope'. This was the
only struct that was not already prefixed.

Moves errors to the transport::error module, but re-exports them within
the matrix module.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>