Add shutdown_all_services method to orchestrator
Adds the `ServiceOrchestrator::shutdown_all_services` method for
shutting down all services managed by the orchestrator. This will be
used by the admin service. Currently, the admin service tracks all
initialized service so it knows which ones to shutdown; this method will
eliminate the need for the admin service to track running services.

Signed-off-by: Logan Seeley <seeley@bitwise.io>