Reorganize orchestrator usage in admin service
Updates how the admin service uses the `ServiceOrchestrator`:
- Share the `ServiceOrchestrator` between `AdminService` and
`AdminServiceShared`
- Replace the `AdminServiceShared::stop_services` method with a call
directly to `ServiceOrchestrator::shutdown_all_services`. This removes
the need for the admin service to track running services.
- Move the `restart_services` method from `AdminServiceShared` to
`AdminService`. This is an important change, because it allows the admin
service to release the lock on `AdminServiceShared` while restarting
services; holding onto the lock caused a deadlock in some cases. To
support this change, the `get_circuits` method has been added to
`AdminServiceShared` for getting circuit definitions in Splinter state.

Signed-off-by: Logan Seeley <seeley@bitwise.io>