Update ConnectionManager Connector not to block on add
Previously, when calling Connector add, the function
would block until the connection had completed authorization.

This commit changes the connector to return after the connection
has been made. If a connection becomes authorized a Connected
Notification is sent to all subscribers. If the connection has
failed authorization a FatalConnectionError notification is sent instead.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>