Update PeerConnector::add_peer_ref not to block
Before add_peer_ref would block until the underlying
peer connection completed authorization. This will be a
problem in the AdminService when all peers must be connected
before the payload can be processed.

This commit changes add_peer_ref to return a peer_ref
before the peer is ready to be used and the peer is set as
Pending. When the underlying connection has completed
authorization, a PeerMangerNotfication::Connected is
sent to subscribers and the peer is set to Connected. If
authorization is failed, a Disconnected notification is sent
instead and the peer stays pending.

add_unidentified_peer requests a connection and the resulting
Connected notification adds the peer as a UnreferencedPeers.

To handle the case where a Pending peer cannot be connected to at
the time of request, a PeerRef is still returned and then the peer
will be retried with an exponential back off. A Pacemaker was added
to drive this retry with the new PeerManagerMessage::RetryPending
message.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>