Fix state_dir handling and reduce use of "storage_location"
This reduces the use of "storage_location" and "location" by replacing
it with state_dir where it was really used as state_dir.

Also pushes state_dir into the daemon code, removing the separate
concept of registry directory and storage location as configurable
options. Registry directory remains largely the same, but circuits.yaml
path is now constructed in the daemon.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>